+++ spec/lib/tag_options/at_remove_spec.rb
@@ -36,6 +36,26 @@ RSpec.describe TagOptions::Hash do
       options.at(:data, :controller).remove!(/sition\z/)
       expect(options.dig(:data, :controller)).to eq("dropdown")
     end
+
+    it "is expected do nothing when no values are resolved on a non-existant root key" do
+      options.at(:nonexistant).remove!(conditional_value: false)
+      expect(options).not_to have_key(:nonexistant)
+    end
+
+    it "is expected do nothing when no values are resolved on a non-existant nested hash" do
+      options.at(:nonexistant, :nested).remove!(conditional_value: false)
+      expect(options).not_to have_key(:nonexistant)
+    end
+
+    it "is expected to remove the key when all values have been removed" do
+      options.at(:class).remove!("ml-1 mr-1")
+      expect(options).not_to have_key(:class)
+    end
+
+    it "is expected to remove nested keys when all values have been removed" do
+      options.at(:data, :controller).remove!("dropdown transition")
+      expect(options).not_to have_key(:data)
+    end
   end
 
   context "#at(as: :style).remove!" do
@@ -47,6 +67,16 @@ RSpec.describe TagOptions::Hash do
     it "is expected to remove a value matching a regular expression" do
       options.at(:style, as: :style).remove!(/color:/)
       expect(options[:style]).to eq("display: none;")
+    end
+
+    it "is expected do nothing when no values are resolved on a non-existant root key" do
+      options.at(:nonexistant, as: :style).remove!("display: none;": false)
+      expect(options).not_to have_key(:nonexistant)
+    end
+
+    it "is expected do nothing when no values are resolved on a non-existant nested hash" do
+      options.at(:nonexistant, :nested, as: :style).remove!("display: none;": false)
+      expect(options).not_to have_key(:nonexistant)
     end
   end
 end